Stabilise Qt startup defaults; enforce strict warnings across targets
Embed qtquickcontrols2.conf (Style=Basic) so app-owned Qt Quick Controls customisation does not produce warnings; add a small platform qtshims module to apply pre-QGuiApplication defaults consistently across vcstream, tst_qml_ui, and fontprobe.

Centralise the warnings policy via vcstream_warnings and apply it repo-wide, including -Wconversion on non-MSVC builds so narrowing/sign-conversion issues are caught consistently.

Fix FontPreviewSafetyCache::Key qHash overload to match Qt 6's size_t-based signature.

+++ modules/ui/fonts/fontpreviewsafetycache.h
@@ -50,7 +50,7 @@ private:
         int pixelSize;
     };
 
-    friend uint qHash( const FontPreviewSafetyCache::Key &k, uint seed );
+    friend size_t qHash( const FontPreviewSafetyCache::Key &k, size_t seed );
     friend bool operator==( const FontPreviewSafetyCache::Key &a, const FontPreviewSafetyCache::Key &b );
 
     static Key makeKey( const QString &family, int pixelSize );
@@ -74,7 +74,7 @@ inline bool operator==( const FontPreviewSafetyCache::Key &a, const FontPreviewS
     return a.pixelSize == b.pixelSize && a.family == b.family;
 }
 
-inline uint qHash( const FontPreviewSafetyCache::Key &k, uint seed = 0 )
+inline size_t qHash( const FontPreviewSafetyCache::Key &k, size_t seed = 0 )
 {
     return qHash( k.family, seed ) ^ ( qHash( k.pixelSize, seed ) + 0x9e3779b9 + ( seed << 6 ) + ( seed >> 2 ) );
 }
